services/requirements.txt
# Services Requirements
# Common dependencies for all trading system services

# Web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client
httpx==0.25.2
requests==2.31.0

# Data processing
pandas==2.1.4
numpy==1.25.2

# Time and date handling
python-dateutil==2.8.2

# Web scraping and parsing
beautifulsoup4==4.12.2
feedparser==6.0.10

# Async support
asyncio-mqtt==0.13.0

# Validation and serialization
pydantic==2.5.0

# Configuration
python-dotenv==1.0.0

# Logging
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2  # For testing

# Development
black==23.11.0
flake8==6.1.0